persistence: fixing issue 5. When xml record of digital document is build, the technicalMetadata element doesn't contain elements format, resolution, compression, color if those elements would have been empty.

+++ xml/src/main/java/cz/nkp/urnnbn/xml/unmarshallers/DigitalInstanceUnmarshaller.java
@@ -24,7 +24,7 @@ public class DigitalInstanceUnmarshaller extends Unmarshaller {
         DigitalInstance result = new DigitalInstance();
         Element root = doc.getRootElement();
         result.setUrl(elementContentOrNull("url", root));
-        result.setLibraryId(Integer.valueOf(elementContentOrNull("digitalLibraryId", root)));
+        result.setLibraryId(Long.valueOf(elementContentOrNull("digitalLibraryId", root)));
         result.setFormat(elementContentOrNull("format", root));
         result.setAccessibility(elementContentOrNull("accessibility", root));
         return result;
